+++ backend/webtoon-file/src/main/java/com/www/file/dto/MainWebtoonDto.java
@@ -1,10 +1,5 @@
 package com.www.file.dto;
 
-import java.time.LocalDateTime;
-
-import com.www.core.auth.entity.Users;
-import com.www.file.dto.WebtoonListDto.WebtoonListDtoBuilder;
-
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,18 +10,18 @@ import lombok.Setter;
 @NoArgsConstructor
 public class MainWebtoonDto {
 	
-	private int idx;
+	private Long idx;
 	private String title;
 	private String author;
 	private String thumbnail;
 	private int genre1;
 	private int genre2;
-	private int hits;
+	private Long hits;
 	private float epRatingAvg;
 	private Integer[] pagelist;
 	
 	@Builder
-	public MainWebtoonDto(int idx, String author, String title, String thumbnail, int genre1, int genre2, float epRatingAvg, int hits) {
+	public MainWebtoonDto(Long idx, String author, String title, String thumbnail, int genre1, int genre2, float epRatingAvg, Long hits) {
 		this.idx = idx;
 		this.title = title;
 		this.author = author;
